+++ server/server.js
@@ -31,10 +31,7 @@ const signup = require("./routes/userAuth/signup");
 const getToken = require("./routes/userAuth/getToken");
 const search = require("./routes/products/searchProducts");
 const address = require("./routes/order/address");
-const cart = require("./routes/cart/cart");
-const addToCart = require("./routes/cart/addToCart");
-const removeFromCart = require("./routes/cart/removeFromCart");
-const updateQuantity = require("./routes/cart/updateQuantity");
+const cart = require("./routes/cart/cart.routes");
 const payment = require("./routes/order/payment");
 const orders = require("./routes/order/orders");
 // const checkout = require("./routes/checkout");
@@ -89,9 +86,6 @@ app.use("/api/admin/product/edit", editProduct);
 app.use("/api/admin/userData", userList);
 app.use("/api/admin/user", userInfo);
 app.use("/api/cart", cart);
-app.use("/api/cart/add", addToCart);
-app.use("/api/cart/remove", removeFromCart);
-app.use("/api/cart/update", updateQuantity);
 app.use("/api/order/confirm", OrderConfirm);
 app.use("/api/admin/orders", adminOrders);
 app.use("/api/categories_list", CategoriesList);
